+++ src/main/java/portablejim/veinminer/VeinMiner.java
@@ -19,7 +19,6 @@ package portablejim.veinminer;
 
 import net.minecraft.block.Block;
 import net.minecraft.command.ServerCommandManager;
-import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
@@ -42,15 +41,15 @@ import net.minecraftforge.fml.common.network.NetworkRegistry;
 import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.oredict.OreDictionary;
 import org.apache.logging.log4j.Logger;
 import portablejim.veinminer.configuration.ConfigurationSettings;
 import portablejim.veinminer.configuration.ConfigurationValues;
 import portablejim.veinminer.configuration.ToolType;
-import portablejim.veinminer.core.InjectedCalls;
+import portablejim.veinminer.core.CoreEvents;
 import portablejim.veinminer.lib.MinerLogger;
 import portablejim.veinminer.lib.ModInfo;
+import portablejim.veinminer.network.PacketChangeMode;
 import portablejim.veinminer.network.PacketClientPresent;
 import portablejim.veinminer.network.PacketMinerActivate;
 import portablejim.veinminer.network.PacketPingClient;
@@ -58,7 +57,7 @@ import portablejim.veinminer.proxy.CommonProxy;
 import portablejim.veinminer.server.MinerCommand;
 import portablejim.veinminer.server.MinerServer;
 import portablejim.veinminer.util.BlockID;
-import portablejim.veinminer.util.Point;
+import portablejim.veinminer.util.PreferredMode;
 
 import java.io.File;
 import java.util.List;
@@ -86,6 +85,7 @@ public class VeinMiner {
     public ConfigurationSettings configurationSettings;
 
     public MinerServer minerServer = null;
+    public int currentMode = PreferredMode.DISABLED;
 
     public Logger logger;
 
@@ -158,6 +158,7 @@ public class VeinMiner {
         proxy.registerClientEvents();
         proxy.registerCommonEvents();
         MinecraftForge.EVENT_BUS.register(this);
+        MinecraftForge.EVENT_BUS.register(new CoreEvents());
     }
 
     public void setupNetworking() {
@@ -165,6 +166,8 @@ public class VeinMiner {
         networkWrapper.registerMessage(PacketPingClient.Handler.class, PacketPingClient.class, 0, Side.CLIENT);
         networkWrapper.registerMessage(PacketClientPresent.Handler.class, PacketClientPresent.class, 1, Side.SERVER);
         networkWrapper.registerMessage(PacketMinerActivate.Handler.class, PacketMinerActivate.class, 2, Side.SERVER);
+        networkWrapper.registerMessage(PacketChangeMode.Handler.class, PacketChangeMode.class, 3, Side.CLIENT);
+        networkWrapper.registerMessage(PacketChangeMode.Handler.class, PacketChangeMode.class, 4, Side.SERVER);
     }
 
     @SuppressWarnings("UnusedDeclaration")
@@ -217,6 +220,8 @@ public class VeinMiner {
             }
         }
         configurationSettings.saveConfigs();
+
+        proxy.registerPostinitCommands();
     }
 
     @SuppressWarnings("UnusedDeclaration")
@@ -275,28 +280,6 @@ public class VeinMiner {
 
                 configurationSettings.addCongruentBlocks(block1, block2);
             }
-        }
-    }
-
-    @SuppressWarnings("unused")
-    @SubscribeEvent
-    public void blockBreakEventFalse(BlockEvent.BreakEvent event) {
-        Point eventPoint = new Point(event.pos.getX(), event.pos.getY(), event.pos.getZ());
-        if(!event.world.isRemote && !event.isCanceled() && event.getPlayer() instanceof EntityPlayerMP && !minerServer.pointIsBlacklisted(eventPoint)) {
-            MinerLogger.debug(String.format("Block Break (False) at %s | %s | %s || Cancel: %s / %s", event.pos.getX(), event.pos.getY(), event.pos.getZ(), !event.isCancelable(), event.isCanceled()));
-            InjectedCalls.blockMined(event.world, (EntityPlayerMP) event.getPlayer(), event.pos, false, new BlockID(event.state));
-            minerServer.removeFromBlacklist(eventPoint);
-        }
-    }
-
-    @SuppressWarnings("unused")
-    @SubscribeEvent
-    public void blockBreakEventTrue(BlockEvent.HarvestDropsEvent event) {
-        Point eventPoint = new Point(event.pos.getX(), event.pos.getY(), event.pos.getZ());
-        if(!event.world.isRemote && !event.isCanceled() && event.harvester instanceof EntityPlayerMP && !minerServer.pointIsBlacklisted(eventPoint)) {
-            MinerLogger.debug(String.format("Block Break (True) at %s | %s | %s || Cancel: %s / %s", event.pos.getX(), event.pos.getY(), event.pos.getZ(), !event.isCancelable(), event.isCanceled()));
-            InjectedCalls.blockMined(event.world, (EntityPlayerMP) event.harvester, event.pos, true, new BlockID(event.state));
-            minerServer.removeFromBlacklist(eventPoint);
         }
     }
 }